remove tracing pipeline dependency and fix tests
