+++ go.mod
@@ -3,7 +3,6 @@ module github.com/kyma-project/serverless-manager
 go 1.19
 
 require (
-	github.com/kyma-project/telemetry-manager v0.0.0-20230726125032-918f324c31d4
 	github.com/onsi/ginkgo/v2 v2.9.5
 	github.com/onsi/gomega v1.27.7
 	github.com/stretchr/testify v1.8.4
@@ -24,6 +23,7 @@ require (
 	github.com/Masterminds/sprig/v3 v3.2.2 // indirect
 	github.com/Masterminds/squirrel v1.5.3 // indirect
 	github.com/asaskevich/govalidator v0.0.0-20210307081110-f21760c49a8d // indirect
+	github.com/benbjohnson/clock v1.3.0 // indirect
 	github.com/chai2010/gettext-go v1.0.2 // indirect
 	github.com/containerd/containerd v1.6.6 // indirect
 	github.com/cyphar/filepath-securejoin v0.2.3 // indirect
